refactor models into modules, clean up docs
