+++ model/src/applications/interaction/kind.rs
