+++ model/src/applications/mod.rs
@@ -1,20 +1,2 @@
-mod command;
-mod command_callback_data;
-mod command_option;
-mod interaction;
-mod interaction_data;
-mod interaction_option_type;
-mod interaction_response;
-mod interaction_type;
-
-pub use command::{Command, CommandData, CommandDataOption};
-pub use command_callback_data::CommandCallbackData;
-pub use command_option::{
-    BaseCommandOptionData, ChoiceCommandOptionData, CommandOption, CommandOptionChoice,
-    CommandOptionType, OptionsCommandOptionData,
-};
-pub use interaction::Interaction;
-pub use interaction_data::InteractionData;
-pub use interaction_option_type::OptionType;
-pub use interaction_response::InteractionResponse;
-pub use interaction_type::InteractionType;
+pub mod command;
+pub mod interaction;
