+++ model/src/gateway/payload/interaction_create.rs
@@ -1,4 +1,4 @@
-use crate::applications::Interaction;
+use crate::applications::interaction::Interaction;
 use serde::{Deserialize, Serialize};
 use std::ops::{Deref, DerefMut};
